Changed favicon & minor CSS changes; removed unnecessary Lazy loading
