+++ public/index.html
@@ -2,12 +2,12 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/fw-favicon.ico" />
+    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
+    <meta name="theme-color" content="#519d87" />
     <meta
       name="description"
-      content="Web site created using create-react-app"
+      content="'Make Food, Not Waste' is a website dedicated to help you reduce your food waste by enabling you to track what ingredients you have in your pantry and search for recipes that will help you use up these items, by turning them into delicious, nutritious meals that fit with your dietary preferences. Whether you're vegan, vegetarian, gluten free, or you eat anything, you'll find recipes to suit you - and to help you Make Food, Not Waste!"
     />
     <link
       rel="apple-touch-icon"
