+++ src/App.css
@@ -1010,8 +1010,8 @@ label {
 }
 
 .activePage {
-  color: #333;
-  background-color: #a5ccb9;
+  color: #519d87;
+  background-color: #f5f5f5;
 }
 
 .mainNav a:first-of-type {
@@ -1019,9 +1019,9 @@ label {
 }
 
 .mainNav a:hover {
-  color: #333;
+  color: #519d87;
   cursor: pointer;
-  background-color: #a5ccb9;
+  background-color: #f5f5f5;
 }
 
 @media screen and (min-width: 540px) {
