+++ src/components/RecipeSearch.jsx
@@ -1,11 +1,10 @@
-import React, { useState, lazy, Suspense } from "react";
+import React, { useState } from "react";
 import Navigation from "./Navigation";
 import SearchBar from "./RecipeSearch/SearchBar";
 import { useSelector } from "react-redux";
 import RecipeTemplate from "./RecipeSearch/RecipeTemplate";
 import LoadingModal from "./Modal/LoadingModal";
-
-const SearchResults = lazy(() => import("./RecipeSearch/SearchResults"));
+import SearchResults from "./RecipeSearch/SearchResults";
 
 const RecipeSearch = () => {
   const [searchTerm, setSearchterm] = useState("");
@@ -28,9 +27,7 @@ const RecipeSearch = () => {
             <RecipeTemplate />
           </div>
         ) : null}
-        <Suspense fallback={<LoadingModal />}>
-          <SearchResults suggestions={suggestions} />
-        </Suspense>
+        <SearchResults suggestions={suggestions} />
       </div>
       {loading ? <LoadingModal /> : null}
     </>
