+++ src/components/RecipeSearch/SearchResults.jsx
@@ -1,13 +1,10 @@
-import React, { useState, lazy, Suspense } from "react";
+import React, { useState } from "react";
 import { useSelector } from "react-redux";
 import NoResults from "./SearchResults/NoResults";
 import ShowResultsButtons from "./SearchResults/ShowResultsButtons";
 import RecipeModal from "./SearchResults/RecipeModal";
 import LoadingModal from "../Modal/LoadingModal";
-
-const SingleRecipeResult = lazy(() =>
-  import("./SearchResults/SingleRecipeResult")
-);
+import SingleRecipeResult from "./SearchResults/SingleRecipeResult";
 
 const SearchResults = (props) => {
   const suggestions = props.suggestions;
@@ -47,27 +44,25 @@ const SearchResults = (props) => {
       <ul className="recipeSearchResults__list">
         {suggestions
           ? suggestions?.map((recipe, index) => {
-              if (index > showMore) return;
+              if (index > showMore) return null;
               return (
-                <Suspense fallback={<LoadingModal />} key={recipe.id}>
-                  <li className="singleRecipe__container">
-                    <SingleRecipeResult
-                      recipe={recipe}
-                      showRecipeMethod={showRecipeMethod}
-                      setShowRecipeMethod={setShowRecipeMethod}
-                      openModal={openModal}
-                      setOpenModal={setOpenModal}
-                      getModalContent={getModalContent}
-                      setIsLoaded={setIsLoaded}
-                    />
-                  </li>
-                </Suspense>
+                <li className="singleRecipe__container" key={recipe.id}>
+                  <SingleRecipeResult
+                    recipe={recipe}
+                    showRecipeMethod={showRecipeMethod}
+                    setShowRecipeMethod={setShowRecipeMethod}
+                    openModal={openModal}
+                    setOpenModal={setOpenModal}
+                    getModalContent={getModalContent}
+                    setIsLoaded={setIsLoaded}
+                  />
+                </li>
               );
             })
           : null}
         {!suggestions && ingredientSearch && ingredientSearch.length > 0
           ? ingredientSearch.map((recipe, index) => {
-              if (index > showMore) return;
+              if (index > showMore) return null;
               else
                 return (
                   <li className="singleRecipe__container" key={recipe.id}>
